get products works again
